Move subsidy_is_correct_test to block::tests
